Refactor reducer
