+++ project/src/components/private-route/private-route.jsx
@@ -25,8 +25,8 @@ PrivateRoute.propTypes = {
   render: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
+const mapStateToProps = ({ USER }) => ({
+  authorizationStatus: USER.authorizationStatus,
 });
 
 export { PrivateRoute };
