+++ project/src/store/user/user.js
@@ -1,0 +1,32 @@
+import { ActionType } from '../action';
+import { AuthorizationStatus } from '../../const';
+
+const initialState = {
+  authorizationStatus: AuthorizationStatus.UNKNOWN,
+  user: null,
+};
+
+const user = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.REQUIRE_AUTH:
+      return {
+        ...state,
+        authorizationStatus: action.payload,
+      };
+    case ActionType.LOG_OUT:
+      return {
+        ...state,
+        authorizationStatus: AuthorizationStatus.NO_AUTH,
+        user: null,
+      };
+    case ActionType.SET_USER:
+      return {
+        ...state,
+        user: action.payload,
+      };
+    default:
+      return state;
+  }
+};
+
+export { user };
